+++ Omise/Resources/CustomerSpecificCardResource.cs
@@ -5,7 +5,7 @@ namespace Omise.Resources
     public class CustomerSpecificCardResource : BaseResource<Card>,
     IListable<Card>,
     IListRetrievable<Card>,
-    IUpdatable<Card, UpdateCardRequest>,
+    IUpdatable<Card, UpdateCardParams>,
     IDestroyable<Card>
     {
         public CustomerSpecificCardResource(IRequester requester, string customerId)
