+++ Omise/Resources/SourceResource.cs
@@ -3,9 +3,9 @@ using Omise.Models;
 
 namespace Omise.Resources
 {
-    public class SourceResource : BaseResource<PaymentSource>,
-        ICreatable<PaymentSource, CreateSourceParams>,
-        IListRetrievable<PaymentSource>
+    public class SourceResource : BaseResource<Source>,
+        ICreatable<Source, CreateSourceParams>,
+        IListRetrievable<Source>
     {
         public SourceResource(IRequester requester)
         : base(requester, Endpoint.Api, "/sources")
